Fix a couple of bugs in the SignUp controller
Upgrading to AutoMapper 2.0 allowed the use of the simpler DynamicMap
API, but this meant that the EmailAddress was no longer being mapped
between the RequiredInfoViewModel and the RegistrationInfo model.

Additionally, the subdomain field was not being set on the
RegistrationInfo model; this is now obtained from the Charity name via a
service call.

Finally, a block of commented-out code has been removed as the method
has been moved to the sign-up service.