Complete the integration test.
Add a SetUp() method to truncate all the tables in the database. Shorten
the test Charity name to make it fit the subdomain field.

The integration tests' database currently has the ConnectionString field
in Charity set to be nullable. This is a conflict with the desired
behaviour, and will need to be updated later when this feature has been
developed.